+++ src/pages/Auth/Login/hooks/entity.ts
@@ -1,0 +1,5 @@
+export interface LoginFormInput {
+  id: string;
+  password: string;
+  isAutoLoginChecked: boolean;
+}
